REPORTS/MAY-MONTH--TASKS--ALL-/Django-React-week-3--until-may-02-/backend/requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
